requirements.txt
Django==1.9.5
psycopg2
gunicorn
djangorestframework==3.3.3
